requirements.txt
PyMuPDF == 1.27.2.2
pandas == 2.3.3
pillow == 12.1.1
python-dotenv == 1.2.2
streamlit